requirements.txt
numpy
numba
torch
opencv-python
pyyaml
pandas
